Calculate biomass from DBH data in fd_inventory()

calc_biomass() uses biomass equations from fd_table_biomass_allometries.csv. Parameters a and b and diameter at breast height (DBH, cm) range for the aboveground biomass sans leaves equation Mass = a * DBH^b for tree species occurring in the UMBS-Flux footprint. See fd_inventory_vignette for further information.

calc_biomass()

Value

A data.frame or tibble of tree biomass calculated using allometries (listed in ?calc_biomass(). Call fd_metadata for field metadata.

Details

Studies used to augment Cooper’s raw data for a species are denoted below.

Species a b DBH range (cm) Acer rubrum a 0.0312 2.7780 1 – 40 Acer saccharum b 0.1693 2.3436 3 – 66 Amelanchier spp.c 0.1630 2.4940 0.3-2.7h Betula papyrifera d 0.0301 2.8387 2 – 49 Fagus grandifolia b 0.1892 2.3097 3 – 66 Pinus resinosa e 0.0526 2.5258 2 – 32 Pinus strobus e 0.0408 2.5735 1 – 32 Populus grandidentata f 0.1387 2.3498 1 – 37 Populus tremuloides b 0.0589 2.6235 3 – 51 Quercus rubra g 0.0398 2.7734 1 – 44

a Crow and Erdmann, 1983, Perala and Alban, 1994, Young et al., 1980. b Young et al., 1980 equations used exclusively, DBH range estimated by Ker-Mikaelian and Korzukhin, 1997. c Perala and Alban, 1994 equations used exclusively. d Perala and Alban, 1994, Schmitt and Grigal, 1981, Young et al., 1980. e Ker, 1980, Perala and Alban, 1994, Young et al., 1980. f Koerper, 1977 raw data collected near UMBS-Flux tower. g Hocker and Early, 1983, Perala and Alban, 1994, Wiant et al., 1977. h Diameter at 15 cm height.

Examples

fortedata::calc_biomass()

#> Warning: These data are unpublished. Please contact Jeff Atkins to ask about using

#> Data citation: ESSD

#> Contact person: Jeff Atkins