LAI values at the subplot scale from littertrap data

LAI is calculated using the equation SLA * mass, where SLA is specific leaf area and mass is leaf litter mass from fd_litter(). Species and site specific values are in fd_sla.csv. Also, see fd_litter_vignette for further explanation, including references, calculation, and possible use cases.

calc_lai()

Value

A data frame of leaf area by year, at the subplot scale.

Examples

calc_lai()

#> # A tibble: 32 x 6
#>    subplot_id replicate  plot subplot  year   lai
#>    <chr>      <chr>     <int> <chr>   <int> <dbl>
#>  1 A01E       A             1 E        2018  4.79
#>  2 A01W       A             1 W        2018  3.82
#>  3 A02E       A             2 E        2018  5.26
#>  4 A02W       A             2 W        2018  4.84
#>  5 A03E       A             3 E        2018  7.29
#>  6 A03W       A             3 W        2018  4.36
#>  7 A04E       A             4 E        2018  6.08
#>  8 A04W       A             4 W        2018  5.44
#>  9 B01E       B             1 E        2018  4.22
#> 10 B01W       B             1 W        2018  4.82
#> # ... with 22 more rows