Canopy structural traits from 2D canopy LiDAR.

Canopy structural traits from 2D canopy LiDAR.

fd_canopy_structure()

Value

A data.frame or tibble of hemispherical camera data. Call fd_metadata for field metadata.

Note

Data were collected by Jeff W. Atkins and Brandon Alveshare using an upward facing, portable canopy lidar unit. The Canopy structural traits were derived using the forestr 1.0.1 package from 2D portable canopy lidar

Examples

fd_canopy_structure()

#> # A tibble: 62 x 33
#>    subplot_id replicate  plot subplot  year mean.height height.2 mean.height.var
#>    <chr>      <chr>     <int> <chr>   <int>       <dbl>    <dbl>           <dbl>
#>  1 D04E       D             4 E        2018        6.09     2.83            7.99
#>  2 D04E       D             4 E        2018        7.72     3.81           14.5 
#>  3 D04W       D             4 W        2018        5.73     2.94            8.65
#>  4 D04W       D             4 W        2018        5.30     3.03            9.19
#>  5 D02E       D             2 E        2018        7.73     3.43           11.8 
#>  6 D02E       D             2 E        2018        9.48     4.94           24.4 
#>  7 D02W       D             2 W        2018        9.74     4.26           18.1 
#>  8 D02W       D             2 W        2018        7.90     3.95           15.6 
#>  9 D01E       D             1 E        2018        9.93     2.48            6.13
#> 10 D01E       D             1 E        2018        8.51     3.68           13.5 
#> # ... with 52 more rows, and 25 more variables: mean.height.rms <dbl>,
#> #   transect.length <int>, mode.el <dbl>, max.el <dbl>, mode.2 <dbl>,
#> #   max.can.ht <dbl>, mean.max.ht <dbl>, mean.vai <dbl>, mean.peak.vai <dbl>,
#> #   deep.gaps <int>, deep.gap.fraction <dbl>, porosity <dbl>, std.std <dbl>,
#> #   mean.std <dbl>, rugosity <dbl>, top.rugosity <dbl>, mean.return.ht <dbl>,
#> #   sd.return.ht <dbl>, sky.fraction <dbl>, cover.fraction <dbl>, max.ht <dbl>,
#> #   scan.density <dbl>, rumple <dbl>, clumping.index <dbl>, enl <dbl>