Summary data for canopy structural data including canopy complexity and leaf area by replicate by year for 2018 to 2020

Summary data for canopy structural data including canopy complexity and leaf area by replicate by year for 2018 to 2020

fd_canopy_structure_summary()

Value

A data.frame or tibble. See "Details" for column descriptions.

Details

The columns are as follows:

replicate (character): Replicate code, extracted from subplot_id.

plot (integer): Plot ID number, extracted from subplot_id.

subplot (character): Subplot code, extracted from subplot_id.

year (integer): Year in which measurement was taken

rugosity (numeric): mean of leaf area index

rugosity_sd (numeric): sd of leaf area index

rugosity_n (integer): number of observations per sample

rugosity_se (numeric): se of leaf area index

vai_mean (numeric): mean of leaf area index

vai_mean_sd (numeric): sd of leaf area index

vai_mean_n (integer): number of observations per sample

vai_mean_se (numeric): se of leaf area index

Note

This summary reports only canopy structural complexity (as canopy rugosity) and leaf area (as VAI, or vegetation area index--akin to PAI or LAI).

Author

Measurements by Jeff Atkins at the University of Michigan Biological Station.

Examples

fd_canopy_structure_summary()

#> # A tibble: 12 x 10
#>    replicate  year rugosity rugosity_sd rugosity_n rugosity_se vai_mean
#>    <chr>     <int>    <dbl>       <dbl>      <int>       <dbl>    <dbl>
#>  1 A          2018    29.6        10.5         129       0.924     6.80
#>  2 A          2019    29.7         9.80        139       0.831     6.00
#>  3 A          2020    24.6        10.6         135       0.911     6.95
#>  4 B          2018    22.4         5.76         70       0.688     7.23
#>  5 B          2019    23.6         6.72         81       0.747     6.52
#>  6 B          2020    22.1         5.52         86       0.595     7.05
#>  7 C          2018    14.0         5.90        109       0.565     6.70
#>  8 C          2019    14.6         4.80         93       0.497     5.94
#>  9 C          2020    14.9         4.17        105       0.407     5.93
#> 10 D          2018     9.08        4.83        121       0.439     5.83
#> 11 D          2019    10.6         2.43         78       0.276     5.26
#> 12 D          2020     8.75        3.86         91       0.405     4.78
#> # ... with 3 more variables: vai_mean_sd <dbl>, vai_mean_n <int>,
#> #   vai_mean_se <dbl>