Ceptometer data.

Ceptometer derived light interception data

fd_ceptometer()

Value

A data.frame or tibble. See "Details" for column descriptions. Call fd_metadata for field metadata.

Details

data on light interception, as faPAR or the fraction of absorbed photosynthetically available radiation, based on the difference between above and below canopy PAR as measured by a Decagon LP-80 handheld ceptometer. Units of a_par, above canopy PAR, and b_par, below canopy PAR, are in units of micromoles per square meter per second while fapar and lai_cept, leaf area index derived from light data, are ratio-based, unitless data. Entries where the notes column has entries may refer to original values where the a_par value was replaced with data from the nearby UMBS AmeriFlux tower PAR sensor due to ceptometer instrument issues with securing clear sky PAR readings. Tower PAR readings were synced in time and then adjusted with the calibration equation A_par = 1.15x - 187.827, where x = the Tower PAR reading. This is an empirically derived equation from several calibration readings found in the supplemental section of Atkins et al. (In Prep)--to be updated when submitted.

Note

Data were collected by Jeff W. Atkins and Brandon Alveshare using a Decagon LP-80 Handheld Ceptometer

Examples

fd_ceptometer()

#> # A tibble: 83 x 10
#>    subplot_id replicate  plot subplot timestamp           a_par b_par fapar
#>    <chr>      <chr>     <int> <chr>   <dttm>              <dbl> <dbl> <dbl>
#>  1 D01W       D             1 W       2018-07-25 10:45:00 1276.  61.8 0.952
#>  2 D03W       D             3 W       2018-07-25 10:56:00 1305.  98.3 0.925
#>  3 D04W       D             4 W       2018-07-25 11:07:00 1305. 345.  0.735
#>  4 D02W       D             2 W       2018-07-25 11:17:00 1367.  58   0.958
#>  5 C04W       C             4 W       2018-07-25 11:24:00 1439.  40.6 0.972
#>  6 C03W       C             3 W       2018-07-25 11:34:00 1383.  43.5 0.969
#>  7 C02W       C             2 W       2018-07-25 11:45:00 1464.  49.5 0.966
#>  8 C01W       C             1 W       2018-07-25 11:52:00 1475. 116   0.921
#>  9 B04W       B             4 W       2018-07-25 13:25:00 1708.  61.7 0.964
#> 10 B03E       B             3 E       2018-07-25 13:36:00 1768. 110.  0.938
#> # ... with 73 more rows, and 2 more variables: lai_cept <dbl>, notes <chr>