Hemispherical camera data.

Hemispherical camera data.

fd_hemi_camera()

Value

A data.frame or tibble. Call fd_metadata for field metadata.

Note

Data were collected by Jeff W. Atkins and Evan Pari using a 24 Megapixel Sony 6000 DSLR Compact 2571 camera (Regent Instruments; Quebec, QU, Canada) with a 180° hemispherical lens. The blue channel of the camera is replaced with a near-infrared channel, which allows direct calculation of NDVI.

Examples

fd_hemi_camera()

#> Warning: `...` is not empty.
#> 
#> We detected these problematic arguments:
#> * `needs_dots`
#> 
#> These dots only exist to allow future extensions and should be empty.
#> Did you misspecify an argument?

#> # A tibble: 1,189 x 11
#>    subplot_id replicate  plot subplot nested_subplot date        ndvi
#>    <chr>      <chr>     <int> <chr>            <int> <date>     <dbl>
#>  1 A01E       A             1 E                    0 2018-07-23 0.796
#>  2 A01E       A             1 E                    1 2018-07-23 0.827
#>  3 A01E       A             1 E                    3 2018-07-23 0.739
#>  4 A01E       A             1 E                    5 2018-07-23 0.847
#>  5 A01E       A             1 E                    7 2018-07-23 0.822
#>  6 A01W       A             1 W                    0 2018-07-23 0.792
#>  7 A01W       A             1 W                    1 2018-07-23 0.862
#>  8 A01W       A             1 W                    3 2018-07-23 0.861
#>  9 A01W       A             1 W                    5 2018-07-23 0.783
#> 10 A01W       A             1 W                    7 2018-07-23 0.786
#> # ... with 1,179 more rows, and 4 more variables: gap_fraction <dbl>,
#> #   openness <dbl>, lai_cam <dbl>, clumping_index <dbl>