Forest Inventory Data

Forest Inventory Data

fd_inventory()

Value

A data.frame or tibble. Call fd_metadata for field metadata.

Details

Forest inventory data for the 32 FoRTE subplots. Includes diameter-at-breast-height (DBH) measurements in units of cm as well as information on health (i.e. see health_status column where D = dead, M = moribund, and L = live) and canopy status (i.e. see canopy_status where OD = overstory dominant, UN = understory, OS = overstory submissive, SA = sapling, and NA is a blank or missing record) and species, not species are identified using the USDA Taxon system (e.g. FAGR is Fagus grandfolia, QURU is Quercus rubra, etc.). The tag column is the tree tag number. These data were originally collected in 2018, pre-disturbance. For stem-girdled tree information, i.e. which trees were targeted for mortality, see fd_mortatlity(). Remeasure for dbh is planned for 2022.

Note

Data were collected by Gough Lab team members Autym Shafer, Catherine McGuigan, and Alexandra Barry in 2018 using a Haglof Postex Inventory Unit with PD II Caliper

Examples

fd_inventory()

#> These data are unpublished. Please contact Jeff Atkins to ask about using

#> Data citation: ESSD

#> Contact person: Jeff Atkins